Closes #5040: Migrate account and push observers
